Move input disposal regression test
